requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==23.1.0
Authlib==1.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==40.0.2
decorator==5.1.1
fastapi==0.95.1
filelock==3.11.0
frozenlist==1.3.3
h11==0.14.0
httptools==0.5.0
huggingface-hub==0.13.4
idna==3.4
multidict==6.0.4
numpy==1.24.2
openai==0.27.4
packaging==23.1
pycparser==2.21
pydantic==1.10.7
python-dotenv==1.0.0
PyYAML==6.0
regex==2023.3.23
requests==2.28.2
sniffio==1.3.0
starlette==0.26.1
tokenizers==0.13.3
tqdm==4.65.0
transformers==4.28.1
typing_extensions==4.5.0
urllib3==1.26.15
uvicorn==0.21.1
uvloop==0.17.0
validators==0.20.0
watchfiles==0.19.0
weaviate-client==3.15.6
websockets==11.0.1
yarl==1.8.2
